try to fix resultsrowview
